tensor stats cache is now a generic cache
